feat(web): add accordion settings layout
