+++ apps/web/components/settings/DataCard.tsx
@@ -1,0 +1,35 @@
+import React from 'react';
+import useAlwaysSD from '../../hooks/useAlwaysSD';
+import { Card } from '../ui/Card';
+
+export function DataCard() {
+  const { alwaysSD, setAlwaysSD } = useAlwaysSD();
+
+  return (
+    <Card title="Data" desc="Playback and usage data.">
+      <Row title="Always play SD (240p)" desc="Reduce data usage on mobile." control={<Switch checked={alwaysSD} onCheckedChange={setAlwaysSD} />} />
+    </Card>
+  );
+}
+
+function Row({ title, desc, control }: { title: string; desc: string; control: React.ReactNode }) {
+  return (
+    <div className="flex items-center justify-between py-3">
+      <div>
+        <div className="font-medium text-sm">{title}</div>
+        <div className="text-xs text-muted-foreground">{desc}</div>
+      </div>
+      {control}
+    </div>
+  );
+}
+
+function Switch({ checked, onCheckedChange }: { checked: boolean; onCheckedChange: (checked: boolean) => void }) {
+  return (
+    <button role="switch" aria-checked={checked} onClick={() => onCheckedChange(!checked)} className={`relative inline-flex h-6 w-10 items-center rounded-full transition-colors ring-offset-2 focus-visible:ring-2 ${checked ? 'bg-accent' : 'bg-foreground/20'}`}>
+      <span className={`inline-block h-4 w-4 transform rounded-full bg-white transition-transform ${checked ? 'translate-x-4' : 'translate-x-1'}`} />
+    </button>
+  );
+}
+
+export default DataCard;
